**What**  
Reduce the time it takes for the UI tests to run:
Locally:
createReferral ~105s -> ~27s
home ~46s -> ~10s

On circleCI (UI tests only):
~200s -> ~122s
**Why**  
To speed up the deployment pipeline